feat: alter to dependencies
