+++ backend/app/schemas/user.py
@@ -47,6 +47,7 @@ class UserLogoutResponse(BaseModel):
 
 
 class UserUpdateRequest(BaseModel):
+    current_password: str
     email: Optional[str] = None
     password: Optional[str] = None
